PB-47: Eventcalendermodal fixes
